Refactor UI components into 'ui' and 'landing-page' folders
Moved and renamed several shared UI components into a dedicated 'ui' directory for better organization and maintainability. Extracted 'JoinCommunity' section into its own component under 'landing-page'. Updated all relevant imports throughout the codebase to reflect these changes.
