+++ app/(admin)/layout.tsx
@@ -1,7 +1,7 @@
 import AdminSidebar from "@/components/layout/admin-sidebar";
 import Footer from "@/components/layout/footer";
 import Header from "@/components/layout/header";
-import Providers from "@/components/providers";
+import Providers from "@/components/layout/providers";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { canAccessDashboard } from "@/lib/user-utils";
 import { getUserByClerkId } from "@/sanity/lib/users/getUserByClerkId";
