+++ components/report/report-form.tsx
@@ -17,7 +17,7 @@ import z from "zod";
 import { Button } from "../ui/button";
 import toast from "react-hot-toast";
 import { useRouter } from "next/navigation";
-import BackButton from "../back-button";
+import BackButton from "../ui/back-button";
 
 type ReportFormProps = {
   content: Post | Comment | User | Topic;
