+++ components/topic/topic-card-skeleton.tsx
@@ -1,9 +1,9 @@
-import Card from "../card";
+import BasicCard from "../basic-card";
 import { Skeleton } from "../ui/skeleton";
 
 const TopicCardSkeleton = () => {
   return (
-    <Card className="flex flex-col gap-3 p-4 w-full">
+    <BasicCard className="flex flex-col gap-3 p-4 w-full">
       <Skeleton className="w-full h-36 rounded-lg" />
 
       <Skeleton className="h-7 w-3/4" />
@@ -17,7 +17,7 @@ const TopicCardSkeleton = () => {
         <Skeleton className="h-4 w-16" />
         <Skeleton className="h-4 w-8" />
       </div>
-    </Card>
+    </BasicCard>
   );
 };
