+++ components/topic/topic-trending-card-skeleton.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Card from "../card";
+import BasicCard from "../basic-card";
 import { Skeleton } from "../ui/skeleton";
 
 const TopicTrendingCardSkeleton = () => {
   return (
-    <Card className="flex flex-row items-center gap-2 p-4 w-full h-20">
+    <BasicCard className="flex flex-row items-center gap-2 p-4 w-full h-20">
       <div className="flex items-center justify-center min-w-[80px] h-16 p-2">
         <Skeleton className="w-12 h-12 rounded-lg" />
       </div>
@@ -12,7 +12,7 @@ const TopicTrendingCardSkeleton = () => {
         <Skeleton className="h-6 w-3/4 mb-2" />
         <Skeleton className="h-4 w-1/3" />
       </div>
-    </Card>
+    </BasicCard>
   );
 };
