+++ components/ui/back-button.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 import { ArrowLeft } from "lucide-react";
 
 type BackButtonProps = {
