+++ components/ui/table-pagination.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { ChevronLeft, ChevronRight } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 import { Table } from "@tanstack/react-table";
 
 type TablePaginationProps<T> = {
