+++ components/user/user-profile-content.tsx
@@ -7,7 +7,7 @@ import { Suspense } from "react";
 import UserProfileStats from "./user-profile-stats";
 import UserPosts from "./user-posts";
 import { SortOrder } from "@/types/filter";
-import SortFilterSelect from "../sort-filter-select";
+import SortFilterSelect from "../ui/sort-filter-select";
 import UserActionsWrapper from "./user-actions-wrapper";
 
 type UserProfileContentProps = {
